ios: moved bridge user event extension to profile
